+++ src/activitypub/views.rs
@@ -45,6 +45,16 @@ pub fn is_activitypub_request(headers: &HeaderMap) -> bool {
             .and_then(|value| value.split(',').next())
             .unwrap_or("");
         return CONTENT_TYPES.contains(&content_type_str);
+    } else {
+        // No Accept header
+        let maybe_user_agent = headers.get("User-Agent")
+            .and_then(|value| value.to_str().ok());
+        if let Some(user_agent) = maybe_user_agent {
+            if user_agent.contains("THIS. IS. GNU social!!!!") {
+                // GNU Social doesn't add Accept header
+                return true;
+            };
+        };
     };
     false
 }