+++ CHANGELOG.md
@@ -20,3 +20,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Fixed post and profile page redirections.
+- Fixed federation with GNU Social.